Module Incr_map_collate

module Collate : sig ... end

module Collated : sig ... end

module Map_list : sig ... end

A representation of a list that provides useful diffing. When constructed from a sorted map in an incremental way, you could reasonably expect that most of the time the diff size will be proportional to the number of elements that are added/removed/changed.

module Compare : sig ... end

module Make (Incr : Incremental.S) : sig ... end